Restructured the app back to the previous structure and changed the manifest so that the intro animation runs before the MainActivity
