+++ Android/islandRush/app/src/main/java/com/quinstedt/islandRush/IntroSplashAnimation.java
@@ -1,0 +1,39 @@
+package com.quinstedt.islandRush;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
+import android.widget.ImageView;
+
+public class IntroSplashAnimation extends AppCompatActivity {
+    ImageView animation;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_intro_splash_animation);
+
+        animation =(ImageView) findViewById(R.id.carImage);
+        Animation animation= AnimationUtils.loadAnimation(getApplicationContext(),R.anim.car_start_animation);
+        this.animation.setAnimation(animation);
+        animation.setAnimationListener(new Animation.AnimationListener() {
+            @Override
+            public void onAnimationStart(Animation animation) {
+
+            }
+
+            @Override
+            public void onAnimationEnd(Animation animation) {
+                finish();
+                startActivity(new Intent(getApplicationContext(),MainActivity.class));
+            }
+
+            @Override
+            public void onAnimationRepeat(Animation animation) {
+
+            }
+        });
+    }
+}
