--- Android/islandRush/app/src/main/res/layout/activity_main_page.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/dynamic_list"
-    tools:context=".MainPage"
-    android:id="@+id/mainPage">
-
-    <EditText
-        android:id="@+id/playerName"
-        android:layout_width="326dp"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:layout_marginStart="50sp"
-        android:layout_marginEnd="50sp"
-        android:hint="Enter players name"
-        android:minHeight="48dp"
-        android:textAlignment="center"
-        android:textColor="@color/white"
-        android:textColorHighlight="#2CDDCF"
-        android:textColorHint="@color/white"
-        android:textColorLink="#08BCD8"
-        android:textSize="20sp" />
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:text="  Enter Race  "
-        android:layout_below="@id/playerName"
-        android:textColor="@color/white"
-        android:textSize="30sp"
-        android:id="@+id/button_enterRace"/>
-
-    <Button
-        android:id="@+id/button_Leaderboard"
-        android:textColor="@color/white"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="LeaderBoard "
-        android:layout_below="@+id/button_enterRace"
-        android:layout_centerInParent="true"
-        android:textSize="30sp" />
-
-</RelativeLayout>
